File renames to reflect new multi_span name.
