+++ include/gsl.h
@@ -21,7 +21,8 @@
 
 #include "gsl_assert.h"  // Ensures/Expects
 #include "gsl_util.h"    // finally()/narrow()/narrow_cast()...
-#include "span.h"           // span, strided_span...
+//#include "span.h"           // span
+#include "multi_span.h"           // multi_span, strided_span...
 #include "string_span.h"    // zstring, string_span, zstring_builder...
 #include <memory>
